+++ packages/terra-application-links/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Changed demo mock background colors to be themable
 
 5.5.0 - (March 12, 2019)
 ------------------