+++ packages/terra-application-links/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Support for icons in application links
 
 5.5.0 - (March 12, 2019)
 ------------------