+++ packages/terra-application-links/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+5.6.0 - (March 20, 2019)
+------------------
 ### Changed
 * Update Jest tests to use terra-enzyme-intl
 * Changed demo mock background colors to be themable